refactor(svelte2tsx): extract TS syntax rewrites into script/type_assertion.rs
Pure move: no logic changes, only `use` additions and `pub(super)` visibility
adjustments. The corresponding unit tests move with the code.
