+++ crates/rsvelte_core/src/svelte2tsx/script/test_support.rs
@@ -1,0 +1,19 @@
+//! Shared helpers for the `script` submodule unit tests.
+
+use super::super::svelte2tsx::{Svelte2TsxOptions, Svelte2TsxResult, svelte2tsx};
+
+pub(super) fn run_svelte2tsx(source: &str) -> Svelte2TsxResult {
+    svelte2tsx(source, Svelte2TsxOptions::default()).expect("svelte2tsx should not fail")
+}
+
+pub(super) fn run_svelte2tsx_ts(source: &str) -> Svelte2TsxResult {
+    svelte2tsx(
+        source,
+        Svelte2TsxOptions {
+            filename: "Component.svelte".to_string(),
+            is_ts_file: true,
+            ..Default::default()
+        },
+    )
+    .expect("svelte2tsx should not fail")
+}
